[WM] fix weird core.async behavior w/ pedestal
Before this change, I was seeing a ClassNotFoundException when it was compiling kehaar.wire-up. Removing pedestal made it go away, but so did removing the try-catch from kehaar.core/go-handler.

I *think* the much-newer version of tools.analyzer.jvm that pedestal brings in (versus the one core.async depends on) was to blame. Hopefully this won't break when core.async updates to a newer version...