[WM] use linear-search pedestal router
This was what pedestal < 0.4.0 used and it is much more flexible and
predictable than the new default prefix-tree router in 0.4.0. The main
advantage of prefix-tree is that it is more performant with lots of
routes. But our very architecture makes it unlikely we'll have enough
routes to matter in any one -http-api component.